requirements.txt
keyboard
pystray
Pillow 